dtype fix #2

+++ modules/arch.py
@@ -4,16 +4,16 @@ import tensorflow as tf
 
 
 class LSTMForgetBlock(tf.keras.layers.Layer):
-    def __init__(self, num_nodes, dtype=tf.float32):
-        super().__init__(name='LSTMForgetBlock', dtype=dtype)
-        self.W_f = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='W_f', use_bias=False)
-        self.U_f = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='U_f')
-        self.W_i = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='W_i', use_bias=False)
-        self.U_i = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='U_i')
-        self.W_o = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='W_o', use_bias=False)
-        self.U_o = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='U_o')
-        self.W_c = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='W_c', use_bias=False)
-        self.U_c = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='U_c')
+    def __init__(self, num_nodes):
+        super().__init__(name='LSTMForgetBlock')
+        self.W_f = tf.keras.layers.Dense(num_nodes, name='W_f', use_bias=False)
+        self.U_f = tf.keras.layers.Dense(num_nodes, name='U_f')
+        self.W_i = tf.keras.layers.Dense(num_nodes, name='W_i', use_bias=False)
+        self.U_i = tf.keras.layers.Dense(num_nodes, name='U_i')
+        self.W_o = tf.keras.layers.Dense(num_nodes, name='W_o', use_bias=False)
+        self.U_o = tf.keras.layers.Dense(num_nodes, name='U_o')
+        self.W_c = tf.keras.layers.Dense(num_nodes, name='W_c', use_bias=False)
+        self.U_c = tf.keras.layers.Dense(num_nodes, name='U_c')
 
     def call(self, x, h, c):
         f = tf.keras.activations.sigmoid(self.W_f(x) + self.U_f(h))
@@ -25,15 +25,15 @@ class LSTMForgetBlock(tf.keras.layers.Layer):
 
 
 class LSTMPeepholeBlock(tf.keras.layers.Layer):
-    def __init__(self, num_nodes, dtype=tf.float32):
-        super().__init__(name='LSTMPeepholeBlock', dtype=dtype)
-        self.W_f = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='W_f', use_bias=False)
-        self.U_f = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='U_f')
-        self.W_i = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='W_i', use_bias=False)
-        self.U_i = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='U_i')
-        self.W_o = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='W_o', use_bias=False)
-        self.U_o = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='U_o')
-        self.W_c = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='W_c')
+    def __init__(self, num_nodes):
+        super().__init__(name='LSTMPeepholeBlock')
+        self.W_f = tf.keras.layers.Dense(num_nodes, name='W_f', use_bias=False)
+        self.U_f = tf.keras.layers.Dense(num_nodes, name='U_f')
+        self.W_i = tf.keras.layers.Dense(num_nodes, name='W_i', use_bias=False)
+        self.U_i = tf.keras.layers.Dense(num_nodes, name='U_i')
+        self.W_o = tf.keras.layers.Dense(num_nodes, name='W_o', use_bias=False)
+        self.U_o = tf.keras.layers.Dense(num_nodes, name='U_o')
+        self.W_c = tf.keras.layers.Dense(num_nodes, name='W_c')
 
     def call(self, x, h, c):
         f = tf.keras.activations.sigmoid(self.W_f(x) + self.U_f(h))
@@ -43,16 +43,16 @@ class LSTMPeepholeBlock(tf.keras.layers.Layer):
         return o*c, c
 
 class DGMBlock(tf.keras.layers.Layer):
-    def __init__(self, num_nodes, dtype=tf.float32):
-        super().__init__(name='DGMBlock', dtype=dtype)
-        self.W_z = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='W_f', use_bias=False)
-        self.U_z = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='U_f')
-        self.W_g = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='W_i', use_bias=False)
-        self.U_g = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='U_i')
-        self.W_r = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='W_o', use_bias=False)
-        self.U_r = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='U_o')
-        self.W_h = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='W_c', use_bias=False)
-        self.U_h = tf.keras.layers.Dense(num_nodes, dtype=dtype, name='U_c')
+    def __init__(self, num_nodes):
+        super().__init__(name='DGMBlock')
+        self.W_z = tf.keras.layers.Dense(num_nodes, name='W_f', use_bias=False)
+        self.U_z = tf.keras.layers.Dense(num_nodes, name='U_f')
+        self.W_g = tf.keras.layers.Dense(num_nodes, name='W_i', use_bias=False)
+        self.U_g = tf.keras.layers.Dense(num_nodes, name='U_i')
+        self.W_r = tf.keras.layers.Dense(num_nodes, name='W_o', use_bias=False)
+        self.U_r = tf.keras.layers.Dense(num_nodes, name='U_o')
+        self.W_h = tf.keras.layers.Dense(num_nodes, name='W_c', use_bias=False)
+        self.U_h = tf.keras.layers.Dense(num_nodes, name='U_c')
 
     def call(self, x, S):
         Z = tf.keras.activations.sigmoid(self.W_z(S) + self.U_z(x))
@@ -70,11 +70,11 @@ class FPForget(tf.keras.models.Model):
         num_layers: number of LSTM layers
     """
     def __init__(self, num_nodes, num_layers, name = 'FPForget', save_path=None):
-        super().__init__(name=name, dtype=tf.float32)
+        super().__init__(name=name)
         self.num_nodes = num_nodes
         self.num_layers = num_layers
-        self.lstm_layers = [LSTMForgetBlock(num_nodes, dtype=self.dtype) for _ in range(num_layers)]
-        self.final_dense = tf.keras.layers.Dense(units=1, activation=tf.keras.activations.exponential, dtype=self.dtype)
+        self.lstm_layers = [LSTMForgetBlock(num_nodes) for _ in range(num_layers)]
+        self.final_dense = tf.keras.layers.Dense(units=1, activation=tf.keras.activations.exponential)
         #self.batch_norm = tf.keras.layers.BatchNormalization(axis=1)
         
         
@@ -82,7 +82,7 @@ class FPForget(tf.keras.models.Model):
     def call(self, *args):
         x = tf.concat(args, axis=1)
         h = tf.zeros_like(x)
-        c = tf.zeros((x.shape[0], self.num_nodes), dtype=self.dtype)
+        c = tf.zeros((x.shape[0], self.num_nodes))
         for i in range(self.num_layers):
             h, c = self.lstm_layers[i](x, h, c)
             #h = self.batch_norm(h)
@@ -100,11 +100,11 @@ class FPPeephole(tf.keras.models.Model):
         num_layers: number of LSTM layers
     """
     def __init__(self, num_nodes, num_layers, name = 'FPPeephole', save_path=None):
-        super().__init__(name=name, dtype=tf.float32)
+        super().__init__(name=name)
         self.num_nodes = num_nodes
         self.num_layers = num_layers
-        self.lstm_layers = [LSTMPeepholeBlock(num_nodes, dtype=self.dtype) for _ in range(num_layers)]
-        self.final_dense = tf.keras.layers.Dense(units=1, activation=tf.keras.activations.exponential, dtype=self.dtype)
+        self.lstm_layers = [LSTMPeepholeBlock(num_nodes) for _ in range(num_layers)]
+        self.final_dense = tf.keras.layers.Dense(units=1, activation=tf.keras.activations.exponential)
         #self.batch_norm = tf.keras.layers.BatchNormalization(axis=1)
         
         
@@ -112,7 +112,7 @@ class FPPeephole(tf.keras.models.Model):
     def call(self, *args):
         x = tf.concat(args, axis=1)
         h = tf.zeros_like(x)
-        c = tf.zeros((x.shape[0], self.num_nodes), dtype=self.dtype)
+        c = tf.zeros((x.shape[0], self.num_nodes))
         for i in range(self.num_layers):
             h, c = self.lstm_layers[i](x, h, c)
             #h = self.batch_norm(h)
@@ -129,12 +129,12 @@ class FPDGM(tf.keras.models.Model):
         num_layers: number of DGM layers
     """
     def __init__(self, num_nodes, num_layers, name = 'FPDGM', save_path=None):
-        super().__init__(name=name, dtype=tf.float32)
+        super().__init__(name=name)
         self.num_nodes = num_nodes
         self.num_layers = num_layers
-        self.initial_dense = tf.keras.layers.Dense(units=num_nodes, activation=tf.keras.activations.tanh, dtype=self.dtype)
-        self.lstm_layers = [DGMBlock(num_nodes, dtype=self.dtype) for _ in range(num_layers)]
-        self.final_dense = tf.keras.layers.Dense(units=1, activation=tf.keras.activations.exponential, dtype=self.dtype)
+        self.initial_dense = tf.keras.layers.Dense(units=num_nodes, activation=tf.keras.activations.tanh)
+        self.lstm_layers = [DGMBlock(num_nodes) for _ in range(num_layers)]
+        self.final_dense = tf.keras.layers.Dense(units=1, activation=tf.keras.activations.exponential)
         #self.batch_norm = tf.keras.layers.BatchNormalization(axis=1)
          
 
@@ -156,12 +156,12 @@ class FPVanilla(tf.keras.models.Model):
         num_layers: number of Vanilla layers
     """
     def __init__(self, num_nodes, num_layers, name = 'FPVanilla', save_path=None):
-        super().__init__(name=name, dtype=tf.float32)
+        super().__init__(name=name)
         self.num_nodes = num_nodes
         self.num_layers = num_layers
-        self.initial_dense = tf.keras.layers.Dense(units=num_nodes, activation=tf.keras.activations.tanh, dtype=self.dtype)
-        self.middle_layers = [tf.keras.layers.Dense(units=num_nodes, activation=tf.keras.activations.tanh, dtype=self.dtype) for _ in range(num_layers)]
-        self.final_dense = tf.keras.layers.Dense(units=1, activation=tf.keras.activations.exponential, dtype=self.dtype)
+        self.initial_dense = tf.keras.layers.Dense(units=num_nodes, activation=tf.keras.activations.tanh)
+        self.middle_layers = [tf.keras.layers.Dense(units=num_nodes, activation=tf.keras.activations.tanh) for _ in range(num_layers)]
+        self.final_dense = tf.keras.layers.Dense(units=1, activation=tf.keras.activations.exponential)
         #self.batch_norm = tf.keras.layers.BatchNormalization(axis=1)
